requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-qdrant
qdrant-client
sentence-transformers
